feat(controllers): implement video controllers
