+++ package.json
@@ -4,6 +4,7 @@
   "description": "rentVideos is a service that helps to rent videos built using node",
   "main": "index.js",
   "scripts": {
+    "start": "nodemon index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
